Separating the Middleware
