+++ Angular/Udemy/SUBHASISH/MEAN Stack/MEAN-STACK-LEARNING/Server/middleware/images.js
@@ -1,0 +1,23 @@
+const multer = require("multer");
+const MIME_TYPE_MAP = {
+  "image/png": "png",
+  "image/jpeg": "jpg",
+  "image/jpg": "jpg",
+};
+const storage = multer.diskStorage({
+  destination: (req, file, cb) => {
+    const isValid = MIME_TYPE_MAP[file.mimetype];
+    let error = new Error("Invalid Mime Type");
+    if (isValid) {
+      error = null;
+    }
+    cb(error, "server/images");
+  },
+  filename: (req, file, cb) => {
+    const name = file.originalname.toLocaleLowerCase().split(" ").join("-");
+    const ext = MIME_TYPE_MAP[file.mimetype];
+    cb(null, name + "-" + Date.now() + "." + ext);
+  },
+});
+
+module.exports = multer({ storage: storage }).single("image");
